django-blog-api/requirements.txt
asgiref==3.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
dj-email-url==1.0.6
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-cache-url==3.4.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
environs==14.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
marshmallow==4.0.0
mysqlclient==2.2.7
packaging==25.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
